requirements.txt
aiohttp==3.1.3
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
async-timeout==2.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.3.0.23
celery==3.1.18
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
kombu==3.0.37
Markdown==3.3.4
multidict==4.7.6
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
yarl==1.6.3
